Adding more tests

+++ pkg/helpers/kvcachehelper/rediscache/rediscache.go
@@ -20,6 +20,10 @@ func NewRedisCache(redisDb redisdb.RedisDB) *RedisCache {
 }
 
 func (c *RedisCache) Set(ctx context.Context, key string, value string) {
+	if key == "" {
+		rlog.Error("Key is empty", nil)
+		return
+	}
 	err := c.redisDb.Set(context.Background(), key, value)
 	if err != nil {
 		rlog.Error(fmt.Sprintf("Error adding value to redis cache by key: %s", key), nil)
@@ -28,6 +32,10 @@ func (c *RedisCache) Set(ctx context.Context, key string, value string) {
 }
 
 func (c *RedisCache) Get(ctx context.Context, key string) (string, bool) {
+	if key == "" {
+		rlog.Error("Key is empty", nil)
+		return "", false
+	}
 	var cacheValue string
 	err := c.redisDb.Get(context.Background(), key, &cacheValue)
 	if err != nil {
@@ -43,6 +51,10 @@ func (c *RedisCache) Get(ctx context.Context, key string) (string, bool) {
 }
 
 func (c *RedisCache) Remove(ctx context.Context, key string) bool {
+	if key == "" {
+		rlog.Error("Key is empty", nil)
+		return false
+	}
 	err := c.redisDb.Delete(context.Background(), key)
 	if err != nil {
 		rlog.Error(fmt.Sprintf("Error deleting value from redis cache by key: %s", key), nil)
